Add actionsBar component with single and multiple items selections
