+++ ui/app/src/components/ItemMenu/ItemMenu.tsx
@@ -17,69 +17,21 @@
 import React, { PropsWithChildren, Suspense } from 'react';
 import { ContextMenuItems, SectionItem } from '../ContextMenu';
 import { Resource } from '../../models/Resource';
-import { DetailedItem, LegacyItem } from '../../models/Item';
+import { DetailedItem } from '../../models/Item';
 import { LookupTable } from '../../models/LookupTable';
 import { useActiveSiteId, useEnv, useLogicResource, usePermissions, useSelection } from '../../utils/hooks';
-import { generateMenuOptions } from './utils';
 import Menu from '@material-ui/core/Menu';
 import { PopoverOrigin, PopoverPosition, PopoverReference } from '@material-ui/core';
-import {
-  CloseChangeContentTypeDialog,
-  closeConfirmDialog,
-  closeCopyDialog,
-  closeCreateFileDialog,
-  closeDeleteDialog,
-  closePublishDialog,
-  closeUploadDialog,
-  showChangeContentTypeDialog,
-  showCodeEditorDialog,
-  showConfirmDialog,
-  showCopyDialog,
-  showCreateFileDialog,
-  showCreateFolderDialog,
-  showDeleteDialog,
-  showDependenciesDialog,
-  showEditDialog,
-  showHistoryDialog,
-  showNewContentDialog,
-  showPreviewDialog,
-  showPublishDialog,
-  showUploadDialog,
-  showWorkflowCancellationDialog
-} from '../../state/actions/dialogs';
-import { fetchWorkflowAffectedItems, getLegacyItemsTree } from '../../services/content';
 import { useDispatch } from 'react-redux';
 import { useIntl } from 'react-intl';
-import { showErrorDialog } from '../../state/reducers/dialogs/error';
-import { batchActions, changeContentType, editTemplate } from '../../state/actions/misc';
-import { fetchItemVersions } from '../../state/reducers/versions';
-import { getRootPath, isValidCutPastePath, withoutIndex } from '../../utils/path';
-import {
-  duplicateAsset,
-  duplicateItem,
-  pasteItem,
-  reloadDetailedItem,
-  setClipBoard,
-  unlockItem
-} from '../../state/actions/content';
-import { popPiece } from '../../utils/string';
+import { getRootPath, isValidCutPastePath } from '../../utils/path';
 import { createStyles, makeStyles } from '@material-ui/core/styles';
-import { translations } from './translations';
 import { rand } from '../Navigation/PathNavigator/utils';
-import {
-  emitSystemEvent,
-  itemCut,
-  showCopyItemSuccessNotification,
-  showCutItemSuccessNotification,
-  showDeleteItemSuccessNotification,
-  showDuplicatedItemSuccessNotification,
-  showEditItemSuccessNotification,
-  showPublishItemSuccessNotification
-} from '../../state/actions/system';
 import Typography from '@material-ui/core/Typography';
 import Skeleton from '@material-ui/lab/Skeleton';
 import { Clipboard } from '../../models/GlobalState';
 import StandardAction from '../../models/StandardAction';
+import { generateSingleItemOptions, itemActionDispatcher } from '../../utils/itemActions';
 
 interface ItemMenuBaseProps {
   path: string;
@@ -161,342 +113,7 @@ export default function ItemMenu(props: ItemMenuProps) {
   });
 
   const onMenuItemClicked = (option: SectionItem) => {
-    switch (option.id) {
-      case 'view': {
-        const path = item.path;
-        const src = `${legacyFormSrc}site=${site}&path=${path}&type=form&readonly=true`;
-        dispatch(showEditDialog({ src }));
-        break;
-      }
-      case 'edit': {
-        const path = item.path;
-        const src = `${legacyFormSrc}site=${site}&path=${path}&type=form`;
-        // TODO: open a embedded form needs the following:
-        // src = `${defaultSrc}site=${site}&path=${embeddedParentPath}&isHidden=true&modelId=${modelId}&type=form`
-
-        fetchWorkflowAffectedItems(site, path).subscribe((items) => {
-          if (items?.length > 0) {
-            dispatch(
-              showWorkflowCancellationDialog({
-                items,
-                onContinue: showEditDialog({
-                  src,
-                  onSaveSuccess: batchActions([showEditItemSuccessNotification(), reloadDetailedItem({ path })])
-                })
-              })
-            );
-          } else {
-            dispatch(
-              showEditDialog({
-                src,
-                onSaveSuccess: batchActions([showEditItemSuccessNotification(), reloadDetailedItem({ path })])
-              })
-            );
-          }
-        });
-        break;
-      }
-      case 'createFolder': {
-        dispatch(
-          showCreateFolderDialog({
-            path: withoutIndex(item.path),
-            allowBraces: item.path.startsWith('/scripts/rest')
-          })
-        );
-        break;
-      }
-      case 'renameFolder': {
-        dispatch(
-          showCreateFolderDialog({
-            path: withoutIndex(item.path),
-            allowBraces: item.path.startsWith('/scripts/rest'),
-            rename: true,
-            value: item.label
-          })
-        );
-        break;
-      }
-      case 'createContent': {
-        dispatch(
-          showNewContentDialog({
-            item,
-            rootPath: getRootPath(item.path),
-            onContentTypeSelected: showEditDialog({})
-          })
-        );
-        break;
-      }
-      case 'delete': {
-        let items = [item];
-        dispatch(
-          showDeleteDialog({
-            items,
-            onSuccess: batchActions([showDeleteItemSuccessNotification(), closeDeleteDialog()])
-          })
-        );
-        break;
-      }
-      case 'changeContentType': {
-        dispatch(
-          showConfirmDialog({
-            title: formatMessage(translations.changeContentType),
-            body: formatMessage(translations.changeContentTypeBody),
-            onCancel: closeConfirmDialog(),
-            onOk: batchActions([
-              closeConfirmDialog(),
-              showChangeContentTypeDialog({
-                item,
-                rootPath: getRootPath(item.path),
-                selectedContentType: item.contentTypeId,
-                onContentTypeSelected: batchActions([
-                  CloseChangeContentTypeDialog(),
-                  changeContentType({ originalContentTypeId: item.contentTypeId, path: item.path })
-                ])
-              })
-            ])
-          })
-        );
-        break;
-      }
-      case 'cut': {
-        dispatch(
-          batchActions([
-            setClipBoard({
-              type: 'CUT',
-              paths: [item.path],
-              sourcePath: item.path
-            }),
-            emitSystemEvent(itemCut({ target: item.path })),
-            showCutItemSuccessNotification()
-          ])
-        );
-        break;
-      }
-      case 'copy': {
-        getLegacyItemsTree(site, item.path, { depth: 1000, order: 'default' }).subscribe(
-          (legacyItem: LegacyItem) => {
-            if (legacyItem.children.length) {
-              dispatch(
-                showCopyDialog({
-                  title: formatMessage(translations.copyDialogTitle),
-                  subtitle: formatMessage(translations.copyDialogSubtitle),
-                  item: legacyItem,
-                  onOk: batchActions([
-                    closeCopyDialog(),
-                    setClipBoard({
-                      type: 'COPY',
-                      sourcePath: item.path
-                    }),
-                    showCopyItemSuccessNotification()
-                  ])
-                })
-              );
-            } else {
-              dispatch(
-                batchActions([
-                  setClipBoard({
-                    type: 'COPY',
-                    paths: [item.path],
-                    sourcePath: item.path
-                  }),
-                  showCopyItemSuccessNotification()
-                ])
-              );
-            }
-          },
-          (response) => {
-            dispatch(
-              showErrorDialog({
-                error: response
-              })
-            );
-          }
-        );
-        break;
-      }
-      case 'paste': {
-        if (clipboard.type === 'CUT') {
-          fetchWorkflowAffectedItems(site, clipboard.sourcePath).subscribe((items) => {
-            if (items?.length > 0) {
-              dispatch(
-                showWorkflowCancellationDialog({
-                  items,
-                  onContinue: pasteItem({ path: item.path })
-                })
-              );
-            } else {
-              dispatch(pasteItem({ path: item.path }));
-            }
-          });
-        } else {
-          dispatch(pasteItem({ path: item.path }));
-        }
-        break;
-      }
-      case 'duplicateAsset': {
-        dispatch(
-          showConfirmDialog({
-            title: formatMessage(translations.duplicate),
-            body: formatMessage(translations.duplicateDialogBody),
-            onCancel: closeConfirmDialog(),
-            onOk: batchActions([
-              closeConfirmDialog(),
-              duplicateAsset({
-                path: item.path,
-                onSuccess: showDuplicatedItemSuccessNotification()
-              })
-            ])
-          })
-        );
-        break;
-      }
-      case 'duplicate': {
-        dispatch(
-          showConfirmDialog({
-            title: formatMessage(translations.duplicate),
-            body: formatMessage(translations.duplicateDialogBody),
-            onCancel: closeConfirmDialog(),
-            onOk: batchActions([
-              closeConfirmDialog(),
-              duplicateItem({
-                path: item.path,
-                onSuccess: showDuplicatedItemSuccessNotification()
-              })
-            ])
-          })
-        );
-        break;
-      }
-      case 'schedule': {
-        dispatch(
-          showPublishDialog({
-            items: [item],
-            scheduling: 'custom',
-            onSuccess: batchActions([
-              showPublishItemSuccessNotification(),
-              reloadDetailedItem({ path: item.path }),
-              closePublishDialog()
-            ])
-          })
-        );
-        break;
-      }
-      case 'publish': {
-        dispatch(
-          showPublishDialog({
-            items: [item],
-            scheduling: 'now',
-            onSuccess: batchActions([
-              showPublishItemSuccessNotification(),
-              reloadDetailedItem({ path: item.path }),
-              closePublishDialog()
-            ])
-          })
-        );
-        break;
-      }
-      case 'history': {
-        dispatch(
-          batchActions([
-            fetchItemVersions({
-              item,
-              rootPath: getRootPath(item.path)
-            }),
-            showHistoryDialog({})
-          ])
-        );
-        break;
-      }
-      case 'dependencies': {
-        dispatch(showDependenciesDialog({ item, rootPath: getRootPath(item.path) }));
-        break;
-      }
-      case 'translation': {
-        break;
-      }
-      case 'editTemplate': {
-        dispatch(editTemplate({ contentTypeId: item.contentTypeId }));
-        break;
-      }
-      case 'editController': {
-        const path = `/scripts/pages/${popPiece(item.contentTypeId, '/')}.groovy`;
-        let src = `${legacyFormSrc}site=${site}&path=${path}&type=controller`;
-
-        fetchWorkflowAffectedItems(site, path).subscribe((items) => {
-          if (items?.length > 0) {
-            dispatch(
-              showWorkflowCancellationDialog({
-                items,
-                onContinue: showCodeEditorDialog({ src })
-              })
-            );
-          } else {
-            dispatch(showCodeEditorDialog({ src }));
-          }
-        });
-        break;
-      }
-      case 'createTemplate': {
-        dispatch(
-          showCreateFileDialog({
-            path: withoutIndex(item.path),
-            type: 'template',
-            onCreated: closeCreateFileDialog()
-          })
-        );
-        break;
-      }
-      case 'createController': {
-        dispatch(
-          showCreateFileDialog({
-            path: withoutIndex(item.path),
-            type: 'controller',
-            onCreated: closeCreateFileDialog()
-          })
-        );
-        break;
-      }
-      case 'codeEditor': {
-        let src = `${legacyFormSrc}site=${site}&path=${encodeURIComponent(item.path)}&type=asset`;
-        dispatch(showCodeEditorDialog({ src }));
-        break;
-      }
-      case 'viewCodeEditor': {
-        let src = `${legacyFormSrc}site=${site}&path=${encodeURIComponent(item.path)}&type=asset&readonly=true`;
-        dispatch(showCodeEditorDialog({ src }));
-        break;
-      }
-      case 'viewImage': {
-        dispatch(
-          showPreviewDialog({
-            type: 'image',
-            title: item.label,
-            url: item.path
-          })
-        );
-        break;
-      }
-      case 'upload': {
-        dispatch(
-          showUploadDialog({
-            path: item.path,
-            site,
-            onClose: closeUploadDialog()
-          })
-        );
-        break;
-      }
-      case 'unlock': {
-        dispatch(
-          unlockItem({
-            path: item.path
-          })
-        );
-        break;
-      }
-      default:
-        break;
-    }
+    itemActionDispatcher(site, item, option, legacyFormSrc, dispatch, formatMessage, clipboard);
     onClose();
   };
   return (
@@ -529,7 +146,7 @@ function ItemMenuUI(props: ItemMenuUIProps) {
     clipboard?.paths.length &&
     getRootPath(clipboard.sourcePath) === getRootPath(item.path) &&
     isValidCutPastePath(item.path, clipboard.sourcePath);
-  const options = generateMenuOptions(item, { hasClipboard, ...permissions });
+  const options = generateSingleItemOptions(item, { hasClipboard, ...permissions });
 
   return <ContextMenuItems classes={classes} sections={options} onMenuItemClicked={onMenuItemClicked} />;
 }
